+++ src/otx/algorithms/classification/utils/__init__.py
@@ -8,10 +8,12 @@ from .cls_utils import (
     get_cls_deploy_config,
     get_cls_inferencer_configuration,
     get_cls_model_api_configuration,
+    get_hierarchical_label_list,
     get_multihead_class_info,
 )
 
 __all__ = [
+    "get_hierarchical_label_list",
     "get_multihead_class_info",
     "get_cls_inferencer_configuration",
     "get_cls_deploy_config",
